requirements.txt
pandas~=1.5.3
Flask~=2.2.2
selenium~=4.8.0
mysql-connector-python~=8.0.32
SQLAlchemy~=2.0.3
python-dateutil~=2.8.2